FIX: Tests in `admin/tests` were not running
Since we want to run them in the core app they've been moved into the
`tests` directory for discourse, and updated to the latest format.
