+++ app/assets/javascripts/discourse/tests/acceptance/admin-install-theme-modal-test.js
