+++ app/assets/javascripts/discourse/tests/integration/widgets/button-test.js
@@ -85,7 +85,7 @@ discourseModule("Integration | Component | Widget | button", function (hooks) {
   });
 
   componentTest("translatedTitle", {
-    template: '{{mount-widget widget="button" args=args}}',
+    template: hbs`{{mount-widget widget="button" args=args}}`,
 
     beforeEach() {
       this.set("args", { label: "topic.create", translatedTitle: "foo bar" });
